Moved extensions from Components to Common
- Extensions aren't exclusively for the components
